improve deduplication of series
